code refactoring: FolderProblemFacadeService 로직을 FolderService에 통합
